% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GO.R
\name{clusterGO}
\alias{clusterGO}
\title{Genes in Cluster}
\usage{
clusterGO(clustNum)
}
\arguments{
\item{clustNum}{infile Path to the input file}
}
\description{
This function takes in the 1. cluster number you are interested 
2. data.val2 and 3. annotation file and tells you how many genes are 
in the cluster of a SOM cluster.
}
