% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SOM_vis.R
\name{clusterVis}
\alias{clusterVis}
\title{Cluster Visualization}
\usage{
clusterVis(clustNum)
}
\arguments{
\item{clustNum}{cluster number you are interested in}
}
\description{
This function makes a ggpot point plot with boxplots overlayed
 based on all six tissue types of that cluster.
}
