% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SOM_vis.R
\name{clusterVis_color}
\alias{clusterVis_color}
\title{Cluster Visualization Color}
\usage{
clusterVis_color(clustNum)
}
\arguments{
\item{clustNum}{cluster number you are interested in}
}
\description{
This function makes a ggplot point plot with boxplots overlayed
 based on all six tissue types of that cluster.
 
 Default palette colors are 
 lcmPaletteColors <- c( "#b3a2ce", "#4753a4","#bf9e71", "#956025","#b2d9a6", "#0f7c3e")
 to match manuscript.
}
