% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/superSOM_vis.R
\name{clusterVis_line_ssom}
\alias{clusterVis_line_ssom}
\title{Cluster Visualization Line for superSOM}
\usage{
clusterVis_line_ssom(clustNum)
}
\description{
Line plot for superSOMs. Not really appropriate, but looks cool.
}
