% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SOM_vis.R
\name{clusterVis_region}
\alias{clusterVis_region}
\title{Cluster Visualization by Region}
\usage{
clusterVis_region(clustNum)
}
\arguments{
\item{clustNum}{cluster number you are interested in}
}
\description{
This function first splits the libraries into tissue types (margin and rachis) 
and logitudinal axis (top, middle, base).  Then, it makes a box plot with 
all points seperated by regions and colored by tissue.
}
