% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/superSOM_vis.R
\name{clusterVis_region_ssom}
\alias{clusterVis_region_ssom}
\title{Cluster Visualization by region for superSOM}
\usage{
clusterVis_region_ssom(clustNum)
}
\description{
Seperated by regions, tiussue and colored by genotype.
}
