% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/num_genes_in_cluster.R
\name{genesInClust}
\alias{genesInClust}
\title{Genes in Cluster}
\usage{
genesInClust(clustNum, data.val2, annotation)
}
\arguments{
\item{clustNum}{The number cluster you want to investigate}
}
\description{
This function takes in the 1. cluster number you are interested 
2. data.val2 and 3. annotation file and tells you how many genes are 
in the cluster of a SOM cluster.
}
