% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colors.R
\docType{data}
\name{lcmPaletteColors}
\alias{lcmPaletteColors}
\title{LCM palette Colors}
\format{An object of class \code{character} of length 6.}
\usage{
lcmPaletteColors
}
\description{
LCM palette Colors
}
\keyword{datasets}
